Rename files to match class name "GraphRAG"
