+++ tests/unit/test_rag.py
@@ -18,7 +18,7 @@ import pytest
 
 from neo4j_genai.exceptions import RagInitializationError, SearchValidationError
 from neo4j_genai.generation.prompts import RagTemplate
-from neo4j_genai.generation.rag import GraphRAG
+from neo4j_genai.generation.graphrag import GraphRAG
 from neo4j_genai.generation.types import RagResultModel
 from neo4j_genai.types import RetrieverResult, RetrieverResultItem
